plane implemented + fixes light inside obj

+++ src/rays/rays.c
@@ -6,7 +6,7 @@
 /*   By: bduval <bduval@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/23 17:08:03 by bduval            #+#    #+#             */
-/*   Updated: 2025/06/04 12:33:47 by bduval           ###   ########.fr       */
+/*   Updated: 2025/06/04 16:01:57 by bduval           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,7 +38,7 @@ int	set_ray(t_ray *ray, t_cam *cam, int x, int y)
 				cam->forward,
 				v_add(v_scale(cam->right, projection.x),
 					v_scale(cam->up, projection.y))));
-	ray->shortest_impact_dist = DBL_MAX;
+	ray->shortest_impact_dist = FLT_MAX;
 	ray->from_cam = 1;
 	return (0);
 }
